examples/xkcd_api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.12.5
    # via httpx
fastapi==0.63.0
    # via xkcd-api (setup.py)
h11==0.12.0
    # via httpcore
httpcore==0.12.3
    # via httpx
httpx==0.16.1
    # via xkcd-api (setup.py)
idna==3.1
    # via rfc3986
jinja2==2.11.3
    # via xkcd-api (setup.py)
markupsafe==1.1.1
    # via jinja2
pydantic==1.7.3
    # via
    #   fastapi
    #   xkcd-api (setup.py)
pyyaml==5.4.1
    # via xkcd-api (setup.py)
rfc3986==1.4.0
    # via httpx
sniffio==1.2.0
    # via
    #   httpcore
    #   httpx
starlette==0.13.6
    # via fastapi
